// tsp-output/server/js/src/generated/models/all/syncly-service.ts
export interface CreatePageRequest {
  ownerName: string;

  editPassword?: string;

  description: string;

  place: string;

  expiresAt: Temporal.Instant;

  candidates: PageCandidateInput[];
}

export interface PageDetail {
  page: PageRecord;

  candidates: PageCandidate[];

  confirmation?: PageConfirmation;

  participants: ParticipantAnswer[];

  availabilitySummary: CandidateAvailabilitySummary[];
}

export interface ErrorResponse {
  code: string;

  message: string;

  details?: string;
}

export interface PageCandidateInput {
  /**
   * 募集日時候補。Asia/Tokyo 基準
   */
  candidateAt: Temporal.Instant;

  /**
   * 表示順
   */
  sortOrder?: number;
}

export interface PageRecord {
  /**
   * 募集ID。ULID
   */
  id: string;

  /**
   * 主催者名
   */
  ownerName: string;

  /**
   * 募集編集用パスワードのハッシュ。パスワードなしの場合は未設定
   */
  editPasswordHash?: string;

  /**
   * 募集概要
   */
  description: string;

  /**
   * 募集場所。オンライン開催URLを含めてもよい
   */
  place: string;

  /**
   * 募集終了日時。Asia/Tokyo 基準
   */
  expiresAt: Temporal.Instant;

  /**
   * 募集状況
   */
  status: PageStatus;

  createdAt: Temporal.Instant;

  updatedAt: Temporal.Instant;

  /**
   * 削除操作日時。DELETED の場合のみ設定する
   */
  deletedAt?: Temporal.Instant;
}

export interface PageCandidate {
  id: string;

  candidateAt: Temporal.Instant;

  sortOrder: number;
}

export interface PageConfirmation {
  candidateId: string;

  candidateAt: Temporal.Instant;

  confirmedAt: Temporal.Instant;
}

export interface ParticipantAnswer {
  id: string;

  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];

  createdAt: Temporal.Instant;

  updatedAt: Temporal.Instant;
}

export interface CandidateAvailabilitySummary {
  candidateId: string;

  candidateAt: Temporal.Instant;

  availableCount: number;
}

export enum PageStatus {
  Open = "OPEN",
  Closed = "CLOSED",
  Expired = "EXPIRED",
  Deleted = "DELETED",
}

export interface UpdatePageRequest {
  ownerName?: string;

  editPassword?: string;

  description?: string;

  place?: string;

  expiresAt?: Temporal.Instant;

  candidates?: PageCandidateInput[];
}

export interface ConfirmPageRequest {
  candidateId: string;
}

export interface CreateParticipantRequest {
  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];
}

export interface UpdateParticipantRequest {
  name?: string;

  canJoinRemotely?: boolean;

  comment?: string;

  availableCandidateIds?: string[];
}

export interface Pages<Context = unknown> {
  /**
   * 募集を作成する
   */
  create(
    ctx: Context,
    body: CreatePageRequest,
  ): Promise<PageDetail | ErrorResponse>;

  /**
   * 募集IDに紐づく募集詳細を取得する
   */
  read(ctx: Context, pageId: string): Promise<PageDetail | ErrorResponse>;

  /**
   * 募集情報と日時候補を更新する
   */
  update(
    ctx: Context,
    pageId: string,
    body: UpdatePageRequest,
  ): Promise<PageDetail | ErrorResponse>;

  /**
   * 募集を論理削除する
   */
  delete(ctx: Context, pageId: string): Promise<void | ErrorResponse>;

  /**
   * 主催者が日時候補を確定し、募集状況をCLOSEDへ更新する
   */
  confirm(
    ctx: Context,
    pageId: string,
    body: ConfirmPageRequest,
  ): Promise<PageDetail | ErrorResponse>;

  /**
   * OPENの募集についてDiscordへリマインドを送信する
   */
  remind(ctx: Context, pageId: string): Promise<void | ErrorResponse>;
}
// ...
export const CreatePageRequest = {
  toJsonObject(input: CreatePageRequest): any {
    return {
      ownerName: input.ownerName,
      editPassword: input.editPassword,
      description: input.description,
      place: input.place,
      expiresAt: input.expiresAt.toString(),
      candidates: input.candidates?.map((item) =>
        PageCandidateInput.toJsonObject(item),
      ),
    };
  },
  fromJsonObject(input: any): CreatePageRequest {
    return {
      ownerName: input.ownerName,
      editPassword: input.editPassword,
      description: input.description,
      place: input.place,
      expiresAt: Temporal.Instant.from(input.expiresAt),
      candidates: input.candidates?.map((item: any) =>
        PageCandidateInput.fromJsonObject(item),
      ),
    };
  },
} as const;

export const PageDetail = {
  toJsonObject(input: PageDetail): any {
    return {
      page: PageRecord.toJsonObject(input.page),
      candidates: input.candidates?.map((item) =>
        PageCandidate.toJsonObject(item),
      ),
      confirmation:
        input.confirmation !== undefined
          ? PageConfirmation.toJsonObject(input.confirmation)
          : undefined,
      participants: input.participants?.map((item) =>
        ParticipantAnswer.toJsonObject(item),
      ),
      availabilitySummary: input.availabilitySummary?.map((item) =>
        CandidateAvailabilitySummary.toJsonObject(item),
      ),
    };
  },
  fromJsonObject(input: any): PageDetail {
    return {
      page: PageRecord.fromJsonObject(input.page),
      candidates: input.candidates?.map((item: any) =>
        PageCandidate.fromJsonObject(item),
      ),
      confirmation:
        input.confirmation !== undefined
          ? PageConfirmation.fromJsonObject(input.confirmation)
          : undefined,
      participants: input.participants?.map((item: any) =>
        ParticipantAnswer.fromJsonObject(item),
      ),
      availabilitySummary: input.availabilitySummary?.map((item: any) =>
        CandidateAvailabilitySummary.fromJsonObject(item),
      ),
    };
  },
} as const;

export const UpdatePageRequest = {
  toJsonObject(input: UpdatePageRequest): any {
    return {
      ownerName: input.ownerName,
      editPassword: input.editPassword,
      description: input.description,
      place: input.place,
      expiresAt:
        input.expiresAt !== undefined ? input.expiresAt.toString() : undefined,
      candidates:
        input.candidates !== undefined
          ? input.candidates?.map((item) =>
              PageCandidateInput.toJsonObject(item),
            )
          : undefined,
    };
  },
  fromJsonObject(input: any): UpdatePageRequest {
    return {
      ownerName: input.ownerName,
      editPassword: input.editPassword,
      description: input.description,
      place: input.place,
      expiresAt:
        input.expiresAt !== undefined
          ? Temporal.Instant.from(input.expiresAt)
          : undefined,
      candidates:
        input.candidates !== undefined
          ? input.candidates?.map((item: any) =>
              PageCandidateInput.fromJsonObject(item),
            )
          : undefined,
    };
  },
} as const;

export const ParticipantAnswer = {
  toJsonObject(input: ParticipantAnswer): any {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: input.createdAt.toString(),
      updatedAt: input.updatedAt.toString(),
    };
  },
  fromJsonObject(input: any): ParticipantAnswer {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: Temporal.Instant.from(input.createdAt),
      updatedAt: Temporal.Instant.from(input.updatedAt),
    };
  },
} as const;

export const PageCandidateInput = {
  toJsonObject(input: PageCandidateInput): any {
    return {
      candidateAt: input.candidateAt.toString(),
      sortOrder: input.sortOrder,
    };
  },
  fromJsonObject(input: any): PageCandidateInput {
    return {
      candidateAt: Temporal.Instant.from(input.candidateAt),
      sortOrder: input.sortOrder,
    };
  },
} as const;

export const PageRecord = {
  toJsonObject(input: PageRecord): any {
    return {
      id: input.id,
      ownerName: input.ownerName,
      editPasswordHash: input.editPasswordHash,
      description: input.description,
      place: input.place,
      expiresAt: input.expiresAt.toString(),
      status: input.status,
      createdAt: input.createdAt.toString(),
      updatedAt: input.updatedAt.toString(),
      deletedAt:
        input.deletedAt !== undefined ? input.deletedAt.toString() : undefined,
    };
  },
  fromJsonObject(input: any): PageRecord {
    return {
      id: input.id,
      ownerName: input.ownerName,
      editPasswordHash: input.editPasswordHash,
      description: input.description,
      place: input.place,
      expiresAt: Temporal.Instant.from(input.expiresAt),
      status: input.status,
      createdAt: Temporal.Instant.from(input.createdAt),
      updatedAt: Temporal.Instant.from(input.updatedAt),
      deletedAt:
        input.deletedAt !== undefined
          ? Temporal.Instant.from(input.deletedAt)
          : undefined,
    };
  },
} as const;

export const PageCandidate = {
  toJsonObject(input: PageCandidate): any {
    return {
      id: input.id,
      candidateAt: input.candidateAt.toString(),
      sortOrder: input.sortOrder,
    };
  },
  fromJsonObject(input: any): PageCandidate {
    return {
      id: input.id,
      candidateAt: Temporal.Instant.from(input.candidateAt),
      sortOrder: input.sortOrder,
    };
  },
} as const;

export const PageConfirmation = {
  toJsonObject(input: PageConfirmation): any {
    return {
      candidateId: input.candidateId,
      candidateAt: input.candidateAt.toString(),
      confirmedAt: input.confirmedAt.toString(),
    };
  },
  fromJsonObject(input: any): PageConfirmation {
    return {
      candidateId: input.candidateId,
      candidateAt: Temporal.Instant.from(input.candidateAt),
      confirmedAt: Temporal.Instant.from(input.confirmedAt),
    };
  },
} as const;

export const CandidateAvailabilitySummary = {
  toJsonObject(input: CandidateAvailabilitySummary): any {
    return {
      candidateId: input.candidateId,
      candidateAt: input.candidateAt.toString(),
      availableCount: input.availableCount,
    };
  },
  fromJsonObject(input: any): CandidateAvailabilitySummary {
    return {
      candidateId: input.candidateId,
      candidateAt: Temporal.Instant.from(input.candidateAt),
      availableCount: input.availableCount,
    };
  },
} as const;